typechecking on every cogs

+++ embedcreator/embedcreator.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import contextlib
 import json
 from copy import deepcopy
-from typing import Any, Optional
+from typing import Any, NoReturn, Optional
 
 import discord
 from discord import Embed
@@ -43,7 +43,9 @@ DEFAULT_EMBED_DESCRIPTION = """
 """.strip()
 
 
-def shorten_by(s: str, /, length: int) -> str:
+def shorten_by(s: str | None, /, length: int) -> str:
+    if not s:
+        return ""
     if len(s) > length:
         return s[: length - 1] + "…"
     return s
@@ -356,7 +358,8 @@ class EmbedFieldRemoverSelect(discord.ui.Select):
 
     async def callback(self, interaction: discord.Interaction):
         self.embed_editor_view.embed.remove_field(int(self.values[0]))
-        await self.embed_editor_view.message.edit(embed=self.embed_editor_view.embed)
+        if self.embed_editor_view.message:
+            await self.embed_editor_view.message.edit(embed=self.embed_editor_view.embed)
         await interaction.response.edit_message(
             content="Field removed.",
             view=None,
@@ -367,11 +370,12 @@ class EmbedFieldRemoverView(discord.ui.View):
     def __init__(self, view: EmbedEditorView):
         super().__init__(timeout=30)
         self.add_item(EmbedFieldRemoverSelect(view))
-        self.message: Optional[discord.Message] = None
+        self.message: discord.Message | None = None
 
     async def on_timeout(self):
-        with contextlib.suppress(discord.HTTPException):
-            await self.message.delete()
+        if self.message:
+            with contextlib.suppress(discord.HTTPException):
+                await self.message.delete()
 
 
 class EmbedEditorView(discord.ui.View):
@@ -388,63 +392,61 @@ class EmbedEditorView(discord.ui.View):
 
     @discord.ui.button(label="Title", style=discord.ButtonStyle.grey)
     async def edit_title_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedTitleModal(self))
 
     @discord.ui.button(label="Description", style=discord.ButtonStyle.grey)
     async def edit_description_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedDescriptionModal(self))
 
     @discord.ui.button(label="Message content", style=discord.ButtonStyle.grey)
     async def edit_message_content_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedMessageContentModal(self))
 
     @discord.ui.button(label="Colour", style=discord.ButtonStyle.grey)
     async def edit_colour_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
-        await interaction.response.send_modal(
-            EmbedColourModal(self, context=self.context)
-        )
+        await interaction.response.send_modal(EmbedColourModal(self, context=self.context))
 
     @discord.ui.button(label="URL", style=discord.ButtonStyle.grey)
     async def edit_url_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedURLModal(self))
 
     @discord.ui.button(label="Image", row=1, style=discord.ButtonStyle.grey)
     async def edit_image_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedImageModal(self))
 
     @discord.ui.button(label="Thumbnail", row=1, style=discord.ButtonStyle.grey)
     async def edit_thumbnail_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedThumbnailModal(self))
 
     @discord.ui.button(label="Author", row=1, style=discord.ButtonStyle.grey)
     async def edit_author_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedAuthorBuilder(self))
 
     @discord.ui.button(label="Footer", row=1, style=discord.ButtonStyle.grey)
     async def edit_footer_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedFooterBuilder(self))
 
     @discord.ui.button(label="Clear", row=1, style=discord.ButtonStyle.red)
     async def clear_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         embed = discord.Embed(description="[…]")
         self.embed = embed
@@ -455,7 +457,7 @@ class EmbedEditorView(discord.ui.View):
         label="Add field", style=discord.ButtonStyle.green, emoji="\U00002795", row=2
     )
     async def add_field_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         await interaction.response.send_modal(EmbedFieldAdder(self))
 
@@ -463,7 +465,7 @@ class EmbedEditorView(discord.ui.View):
         label="Remove field", style=discord.ButtonStyle.red, emoji="\U00002796", row=2
     )
     async def remove_field_button(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         if not self.embed.fields:
             await interaction.response.send_message(
@@ -478,7 +480,7 @@ class EmbedEditorView(discord.ui.View):
         label="Get Python", style=discord.ButtonStyle.blurple, emoji="\U0001f40d", row=3
     )
     async def get_python(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         embed = self.embed
         text = "embed = discord.Embed("
@@ -507,22 +509,14 @@ class EmbedEditorView(discord.ui.View):
             for attr in ["name", "url", "icon_url"]:
                 if gattr := getattr(embed.author, attr):
                     attrs[attr] = gattr
-            text += (
-                "embed.set_author("
-                + ", ".join(f"{k}={v!r}" for k, v in attrs.items())
-                + ")\n"
-            )
+            text += "embed.set_author(" + ", ".join(f"{k}={v!r}" for k, v in attrs.items()) + ")\n"
 
         if embed.footer:
             attrs = {}
             for attr in ["text", "icon_url"]:
                 if gattr := getattr(embed.footer, attr):
                     attrs[attr] = gattr
-            text += (
-                "embed.set_footer("
-                + ", ".join(f"{k}={v!r}" for k, v in attrs.items())
-                + ")\n"
-            )
+            text += "embed.set_footer(" + ", ".join(f"{k}={v!r}" for k, v in attrs.items()) + ")\n"
 
         if not text.endswith("\n\n"):
             text += "\n"
@@ -555,7 +549,7 @@ class EmbedEditorView(discord.ui.View):
         label="Get JSON", style=discord.ButtonStyle.blurple, emoji=JSON_EMOJI, row=3
     )
     async def get_json(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
         text = json.dumps(self.embed.to_dict(), indent=4)
         if len(text) > 1990:
@@ -576,11 +570,9 @@ class EmbedEditorView(discord.ui.View):
         row=3,
     )
     async def replace_json(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
-        await interaction.response.send_modal(
-            EmbedDictionaryUpdater(self, replace=True)
-        )
+        await interaction.response.send_modal(EmbedDictionaryUpdater(self, replace=True))
 
     @discord.ui.button(
         label="Update JSON",
@@ -589,11 +581,9 @@ class EmbedEditorView(discord.ui.View):
         row=3,
     )
     async def update_json(
-        self, interaction: discord.Interaction, button: discord.ui.Button
+        self, interaction: discord.Interaction, button: discord.ui.Button["EmbedEditorView"]
     ):
-        await interaction.response.send_modal(
-            EmbedDictionaryUpdater(self, replace=False)
-        )
+        await interaction.response.send_modal(EmbedDictionaryUpdater(self, replace=False))
 
     @discord.ui.select(
         cls=discord.ui.ChannelSelect,
@@ -609,14 +599,26 @@ class EmbedEditorView(discord.ui.View):
         row=4,
     )
     async def send(
-        self, interaction: discord.Interaction, select: discord.ui.ChannelSelect
+        self, interaction: discord.Interaction, select: discord.ui.ChannelSelect["EmbedEditorView"]
     ):
+        if not interaction.guild:
+            return await interaction.response.send_message(
+                "This command can only be used in a server.", ephemeral=True
+            )
         channel = interaction.guild.get_channel(select.values[0].id)
+        if not channel:
+            return await interaction.response.send_message("Channel not found.", ephemeral=True)
         if not channel.permissions_for(interaction.guild.me).send_messages:
             return await interaction.response.send_message(
                 f"I do not have permissions to post in {channel.mention}.",
                 ephemeral=True,
             )
+        if not isinstance(channel, (discord.TextChannel, discord.Thread)):
+            return await interaction.response.send_message(
+                f"Cannot send messages to {channel.mention}.", ephemeral=True
+            )
+        if not isinstance(interaction.user, discord.Member):
+            return
         if not channel.permissions_for(interaction.user).send_messages:
             return await interaction.response.send_message(
                 f"You do not have permissions to post in {channel.mention}.",
@@ -638,7 +640,8 @@ class EmbedEditorView(discord.ui.View):
             )
 
     async def on_timeout(self):
-        await self.message.edit(view=None)
+        if self.message:
+            await self.message.edit(view=None)
 
     async def modify_embed(self, modal: ModalBase, interaction: discord.Interaction):
         previous_embed = deepcopy(self.embed)
@@ -647,16 +650,17 @@ class EmbedEditorView(discord.ui.View):
         except ValueError as exc:
             return await interaction.response.send_message(f"An error occured: {exc}")
         try:
-            await interaction.message.edit(embed=self.embed, content=self.content)
+            if interaction.message:
+                await interaction.message.edit(embed=self.embed, content=self.content)
         except discord.HTTPException as exc:
-            exc = exc.text.replace("embeds.0.", "embed ")
-            if "maximum size of 6000" in exc:
+            exc_text = exc.text.replace("embeds.0.", "embed ")
+            if "maximum size of 6000" in exc_text:
                 return await interaction.response.send_message(
                     "Sorry, the embed limit has exceeded 6000 characters, which is the maximum size. Your change could not be made.",
                     ephemeral=True,
                 )
             await interaction.response.send_message(
-                f"A HTTP error occured whilst making modifications to the embed:\n{box(exc)}\n"
+                f"A HTTP error occured whilst making modifications to the embed:\n{box(exc_text)}\n"
             )
             self.embed = previous_embed
         else:
@@ -696,28 +700,28 @@ class EmbedArgsConverter(FlagConverter):
         name="source", default=None, converter=commands.MessageConverter
     )
 
-    embed_settable_attributes: tuple[str] = (
+    embed_settable_attributes: tuple[str, str, str, str] = (
         "title",
         "description",
         "colour",
         "url",
     )
 
-    def author_kwargs(self):
-        d = {}
+    def author_kwargs(self) -> dict[str, str]:
+        d: dict[str, str] = {}
         for attr in ("name", "url", "icon_url"):
             if (gattr := getattr(self, f"author_{attr}")) is not None:
                 d[attr] = gattr
         return d
 
-    def footer_kwargs(self):
-        d = {}
+    def footer_kwargs(self) -> dict[str, str]:
+        d: dict[str, str] = {}
         for attr in ("text", "icon_url"):
             if (gattr := getattr(self, f"footer_{attr}")) is not None:
                 d[attr] = gattr
         return d
 
-    def to_dict(self):
+    def to_dict(self) -> dict[str, Any]:
         return {
             "type": "rich",
             "title": self.title,
@@ -737,7 +741,8 @@ class EmbedArgsConverter(FlagConverter):
             },
         }
 
-    async def convert(self, ctx: commands.Context, argument: str):
+    @classmethod
+    async def convert(cls, ctx: commands.Context, argument: str):  # type: ignore[override]
         try:
             return await super().convert(ctx, argument)
         except commands.BadFlagArgument as e:
@@ -767,11 +772,12 @@ class EmbedCreator(commands.Cog):
         context = super().format_help_for_context(ctx)
         return f"{context}\n\nAuthor: {self.__author__}\nVersion: {self.__version__}"
 
-    async def red_delete_data_for_user(self, **kwargs):
-        return
+    async def red_delete_data_for_user(self, **kwargs: Any) -> NoReturn:
+        """Nothing to delete."""
+        raise NotImplementedError
 
-    @commands.command(aliases=["ecreate"])
     @commands.mod()
+    @commands.command(aliases=["ecreate"])
     async def embedcreate(self, ctx: commands.Context, *, options: EmbedArgsConverter):
         """Create an embed.
 
@@ -812,7 +818,6 @@ class EmbedCreator(commands.Cog):
         if options.thumbnail != flags["thumbnail"].default:
             embed.set_thumbnail(url=options.thumbnail)
         if kwargs := options.author_kwargs():
-            print(kwargs)
             embed.set_author(**kwargs)
         if kwargs := options.footer_kwargs():
             embed.set_footer(**kwargs)
@@ -820,9 +825,7 @@ class EmbedCreator(commands.Cog):
         try:
             if options.builder:
                 view.embed = embed
-                view.message = await ctx.send(
-                    view=view, embed=embed, content=options.content
-                )
+                view.message = await ctx.send(view=view, embed=embed, content=options.content)
             else:
                 await ctx.send(embed=embed, content=options.content)
         except discord.HTTPException as exc:
