+++ higherorlower/higherorlower.py
@@ -4,6 +4,7 @@ import contextlib
 import io
 import random
 import typing
+from typing import Any, NoReturn
 
 import discord
 from PIL import Image, ImageDraw, ImageFont
@@ -24,20 +25,25 @@ class HigherOrLowerView(discord.ui.View):
         super().__init__(timeout=30)
 
     @discord.ui.button(label="Higher", style=discord.ButtonStyle.green)
-    async def higher(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def higher(
+        self, interaction: discord.Interaction, button: discord.ui.Button[typing.Self]
+    ):
         self.direction = "higher"
         self.stop()
         await interaction.response.defer()
 
     @discord.ui.button(label="Lower", style=discord.ButtonStyle.red)
-    async def lower(self, interaction: discord.Interaction, button: discord.ui.Button):
+    async def lower(
+        self, interaction: discord.Interaction, button: discord.ui.Button[typing.Self]
+    ):
         self.direction = "lower"
         self.stop()
         await interaction.response.defer()
 
     async def on_timeout(self):
         with contextlib.suppress(discord.HTTPException):
-            await self.message.delete()
+            if self.message is not None:
+                await self.message.delete()
 
 
 class HigherOrLowerSession:
@@ -75,9 +81,7 @@ class HigherOrLowerSession:
         )
 
     def create_image(self) -> discord.File:
-        image = Image.new(
-            "RGBA", (250 * self.size[0], 363 * self.size[1]), self.table_colour
-        )
+        image = Image.new("RGBA", (250 * self.size[0], 363 * self.size[1]), self.table_colour)
         x = 0
         y = 0
         for index, img in enumerate(self.images[: self.progress + 1], 1):
@@ -107,9 +111,7 @@ class HigherOrLowerSession:
         image = Image.new("RGBA", (500, 363), self.table_colour)
 
         if self.ended:
-            card = self.images[self.progress].rotate(
-                -8, resample=Resampling.BICUBIC, expand=True
-            )
+            card = self.images[self.progress].rotate(-8, resample=Resampling.BICUBIC, expand=True)
             image.paste(card, (200, 0), card)
             image.paste(card := self.images[self.progress - 1], (0, 0), card)
 
@@ -128,12 +130,7 @@ class HigherOrLowerSession:
     def evaluate(self, guess: HigherLower):
         current = self.indexes[self.progress]
         following = self.indexes[self.progress + 1]
-        if (
-            current < following
-            and guess == "higher"
-            or current > following
-            and guess == "lower"
-        ):
+        if current < following and guess == "higher" or current > following and guess == "lower":
             return True
         if current == following and self.equal_survives:
             return True
@@ -160,23 +157,21 @@ class HigherOrLowerSession:
         embed.set_footer(text=f"Winning odds: 1/{odds} ({round(1 / odds * 100, 3)}%)")
 
         view = HigherOrLowerView()
-        view.message = message = await ctx.send(
-            embed=embed, files=self.get_files(), view=view
-        )
+        view.message = message = await ctx.send(embed=embed, files=self.get_files(), view=view)
 
         while self.progress != self.size[0] * self.size[1] - 1:
             await view.wait()
 
             if view.timed_out:
                 return
+            if view.direction is None:
+                return
 
             if not self.evaluate(view.direction):
                 embed.colour = discord.Colour.red()
                 embed.set_footer(text="Unlucky, you lose.")
                 self.end_game()
-                return await message.edit(
-                    embed=embed, attachments=self.get_files(), view=None
-                )
+                return await message.edit(embed=embed, attachments=self.get_files(), view=None)
 
             self.progress += 1
 
@@ -185,9 +180,7 @@ class HigherOrLowerSession:
 
         self.won = True
         embed.set_footer(text="You won! 🎉")
-        await message.edit(
-            embed=embed, attachments=self.get_files(exclude_thumb=True), view=None
-        )
+        await message.edit(embed=embed, attachments=self.get_files(exclude_thumb=True), view=None)
 
 
 class HigherOrLower(commands.Cog):
@@ -212,14 +205,17 @@ class HigherOrLower(commands.Cog):
         context = super().format_help_for_context(ctx)
         return f"{context}\n\nAuthor: {self.__author__}\nVersion: {self.__version__}"
 
-    async def red_delete_data_for_user(self, **kwargs):
-        return
+    async def red_delete_data_for_user(self, **kwargs: Any) -> NoReturn:
+        """Nothing to delete."""
+        raise NotImplementedError
 
     @commands.command(aliases=["hol"])
     @commands.guild_only()
     @commands.cooldown(1, 2, commands.BucketType.user)
     async def higherorlower(self, ctx: commands.Context):
         """Play Higher Or Lower!"""
+        if ctx.guild is None:
+            return
         table_colour = await self.config.user(ctx.author).table_colour()
         guild_config = await self.config.guild(ctx.guild).all()
         session = HigherOrLowerSession(
@@ -235,10 +231,11 @@ class HigherOrLower(commands.Cog):
             if session.won:
                 payout = guild_config["payout"]
                 credit_name = await bank.get_currency_name(ctx.guild)
-                await bank.deposit_credits(ctx.author, payout)
-                await ctx.send(
-                    f"Congratulations {ctx.author.mention}, you win {payout} {credit_name}!"
-                )
+                if isinstance(ctx.author, discord.Member):
+                    await bank.deposit_credits(ctx.author, payout)
+                    await ctx.send(
+                        f"Congratulations {ctx.author.mention}, you win {payout} {credit_name}!"
+                    )
 
     @commands.group()
     async def holset(self, ctx: commands.Context):
@@ -249,9 +246,7 @@ class HigherOrLower(commands.Cog):
         """Set the colour of your table used in games."""
         await self.config.user(ctx.author).table_colour.set(colour.to_rgb())
         if await ctx.embed_requested():
-            await ctx.send(
-                embed=discord.Embed(title="Table colour set!", colour=colour)
-            )
+            await ctx.send(embed=discord.Embed(title="Table colour set!", colour=colour))
         else:
             await ctx.send("Table colour set!")
 
@@ -260,6 +255,8 @@ class HigherOrLower(commands.Cog):
     @holset.command(name="payout")
     async def holset_payout(self, ctx: commands.Context, payout: commands.positive_int):
         """Mods only - Set the win payout for this guild."""
+        if ctx.guild is None:
+            return
         await self.config.guild(ctx.guild).payout.set(payout)
         credit_name = await bank.get_currency_name(ctx.guild)
         await ctx.send(f"Payout set to {payout} {credit_name}")
@@ -269,6 +266,8 @@ class HigherOrLower(commands.Cog):
     @holset.command(name="acehigh")
     async def holset_setace(self, ctx: commands.Context, ace_is_high: bool):
         """Mods only - Set whether ace is considered high (14)."""
+        if ctx.guild is None:
+            return
         await self.config.guild(ctx.guild).ace_high.set(ace_is_high)
         await ctx.send(
             f"Ace will now be considered as {'high (14)' if ace_is_high else 'low (1)'}"
@@ -279,6 +278,8 @@ class HigherOrLower(commands.Cog):
     @holset.command(name="equalsurvives")
     async def holset_equalsurvives(self, ctx: commands.Context, survives: bool):
         """Mods only - Set whether players survive on an equal card."""
+        if ctx.guild is None:
+            return
         await self.config.guild(ctx.guild).equal_survives.set(survives)
         grammar = "now" if survives else "now **not**"
         await ctx.send(f"Equal cards will {grammar} be considered a lifeline.")
@@ -288,6 +289,8 @@ class HigherOrLower(commands.Cog):
     @holset.command(name="rotatedstyle")
     async def holset_rotatedstyle(self, ctx: commands.Context, use: bool):
         """Mods only - Set whether placed cards on the table use rotated style."""
+        if ctx.guild is None:
+            return
         await self.config.guild(ctx.guild).rotated_style.set(use)
         grammar = "now" if use else "now **not**"
         await ctx.send(f"Placed cards will {grammar} use rotated style.")
@@ -302,6 +305,8 @@ class HigherOrLower(commands.Cog):
         y: commands.Range[int, 1, 10],
     ):
         """Mods only - Set the grid size."""
+        if ctx.guild is None:
+            return
         if x * y < 4:
             return await ctx.send("Grid size is too small.")
         if x * y > 52:
@@ -313,8 +318,10 @@ class HigherOrLower(commands.Cog):
     @holset.command(name="showsettings", aliases=["settings"])
     async def holset_showsettings(self, ctx: commands.Context):
         """Mods only - See the current settings for Higher Or Lower."""
+        if ctx.guild is None:
+            return
         message = f"# Higher Or Lower settings\nYour table colour: {discord.Colour.from_rgb(*(await self.config.user(ctx.author).table_colour()))}\n"
-        if await self.bot.is_mod(ctx.author):
+        if isinstance(ctx.author, discord.Member) and await self.bot.is_mod(ctx.author):
             config = await self.config.guild(ctx.guild).all()
             for key, value in config.items():
                 if key == "size":
